Add shared confirm dialog and hover to organization-list (#35)

+++ src/app/components/organization-list/organization-list.component.ts
@@ -4,6 +4,7 @@ import {Router} from '@angular/router';
 import {Organization} from '../../model/Organization';
 import {TaskService} from '../../services/task.service';
 import {Task} from '../../model/Task';
+import {DialogService} from '../../services/dialog.service';
 
 @Component({
   selector: 'app-organization-list',
@@ -14,7 +15,8 @@ export class OrganizationListComponent implements OnInit {
 
   organizations: Organization[];
 
-  constructor(private organizationService: OrganizationService, private taskService: TaskService, private router: Router) {
+  constructor(private organizationService: OrganizationService, private taskService: TaskService, private router: Router,
+              private dialogService: DialogService) {
 
   }
 
@@ -33,7 +35,7 @@ export class OrganizationListComponent implements OnInit {
         tasks = data;
         item.numberDone = tasks.filter(task => !task.taskStatus.localeCompare('Done')).length;
         item.numberInProgress = tasks.filter(task => !task.taskStatus.localeCompare('InProgress')).length;
-        item.numberToDo = tasks.filter(task => !task.taskStatus.localeCompare('ToDo')).length;
+        item.numberToDo = tasks.filter(task => !task.taskStatus.localeCompare('Todo')).length;
       });
     });
   }
@@ -43,9 +45,14 @@ export class OrganizationListComponent implements OnInit {
   }
 
   onClickDeleteOrganization(organization: Organization) {
-    console.log(organization);
-    this.organizationService.deleteOrganization(organization.id).subscribe(() => {
-      this.organizations = this.organizations.filter(item => item !== organization);
+    this.dialogService.openConfirmDialog('Are you sure to delete this organization ?')
+      .afterClosed().subscribe(res => {
+      if (res) {
+        console.log(organization);
+        this.organizationService.deleteOrganization(organization.id).subscribe(() => {
+          this.organizations = this.organizations.filter(item => item !== organization);
+        });
+      }
     });
   }
